refactored more spaghetti custom auth. Removed states governing signining in for no reason.
Now to show tasks needs connected and authenticated.
If somebody is not authenticated but connected, wallet should bring them directly to SIWE, not set a flag, and then let the useState flag update and trigger the sign in.
If somebody just came back into focus and the session is expired, a message should show that they should sign in again.